+++ internal/storage/storage.go
@@ -22,7 +22,7 @@ type CommentStorage interface {
 	CreateComment(ctx context.Context, content string, postID int64, parentID *int64) (int64, error)
 	GetCommentByID(ctx context.Context, id int64) (models.Comment, error)
 	GetCommentsByParentID(ctx context.Context, postID int64) ([]models.Comment, error)
-	GetCommentsByPostID(ctx context.Context, postID int64, limit *int64, offset *int64) ([]models.Comment, error)
+	GetCommentsByPostID(ctx context.Context, postID int64, limit *int32, offset *int32) ([]models.Comment, error)
 	DeleteComment(ctx context.Context, id int64) error
 	DeleteCommentsByPostID(ctx context.Context, id int64) error
 }
